トップページの変更と、お気に入りと履歴の調整
