+++ components/Navigation.vue
@@ -1,6 +1,6 @@
 <template>
   <v-bottom-navigation app fixed color="pink lighten-2">
-    <v-btn to="/nurseries/">
+    <v-btn to="/">
       <span>リスト表示</span>
       <v-icon>list</v-icon>
     </v-btn>
