+++ components/nurseries/Detail.vue
@@ -25,16 +25,16 @@
       </v-list-item-content>
     </v-list-item>
     <v-list-item v-if="item.nursery.facility.hasParkingLot">
-      <v-list-content>
+      <v-list-item-content>
         <v-list-item-subtitle>駐車場台数</v-list-item-subtitle>
         <v-list-item-title>{{ item.nursery.facility.numberOfParkingLot }}台</v-list-item-title>
-      </v-list-content>
+      </v-list-item-content>
     </v-list-item>
     <v-list-item v-if="item.nursery.facility.remarksFacility">
-      <v-list-content>
+      <v-list-item-content>
         <v-list-item-subtitle>備考</v-list-item-subtitle>
         <v-list-item-title>{{ item.nursery.facility.remarksFacility }}</v-list-item-title>
-      </v-list-content>
+      </v-list-item-content>
     </v-list-item>
   </v-list>
 </template>
