+++ store/mutation-types.js
@@ -1,2 +1,4 @@
 export const REGISTER_FAVORITE = "favorite/REGISTER";
 export const UNREGISTER_FAVORITE = "favorite/UNREGISTER";
+
+export const ADD_HISTORY = "recent/ADD";
